+++ packages/react-sdk/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './Avatar';
+export * from './BackgroundFilters';
 export * from './Button';
 export * from './CallControls';
 export * from './CallParticipantsList';