+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.4.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.7...@stream-io/video-react-bindings-0.4.8) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.8`
+
+### Features
+
+* **react:** Support for Background Filters and Background Blurring ([#1283](https://github.com/GetStream/stream-video-js/issues/1283)) ([f790ee7](https://github.com/GetStream/stream-video-js/commit/f790ee78c20fb0f5266e429a777d8bb7ef158c83)), closes [#1271](https://github.com/GetStream/stream-video-js/issues/1271) [#1276](https://github.com/GetStream/stream-video-js/issues/1276)
+
 ### [0.4.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.6...@stream-io/video-react-bindings-0.4.7) (2024-04-02)
 
 ### Dependency Updates